requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
Pillow>=10.0,<11.0
gunicorn>=20.1,<21.0
drf-spectacular>=0.26,<0.27
django-storages>=1.13,<1.14
boto3>=1.34,<1.35